moved bot and schedule to appropriate modules
