+++ bot/__init__.py
@@ -1,0 +1,1 @@
+from .bot import send_question, start_bot
