+++ main.py
@@ -4,9 +4,9 @@ import math
 from models import db_session
 from bot import start_bot, send_question
 import schedule
-from tools import Settings
-from web import app as web
-from web import socketio
+from tools import Settings, WeekDays
+from web.web import app as web
+from web.web import socketio
 
 # Environment variables
 # ADMIN_PASSWD: password for web panel
@@ -18,7 +18,7 @@ default_settings = {"tg_pin": "32266",
                     "from_time": datetime.time(7),
                     "to_time": datetime.time(20),
                     "order": 1,
-                    "week_days": [schedule.WeekDays(d) for d in range(7)],
+                    "week_days": [WeekDays(d) for d in range(7)],
                     "distribution_function": math.exp,
                     "repetition_amount": 6,
                     }
